update with changes from feedback
